+++ docker/core/mkwebaxum/src/mk_lib_database_media_images.rs
@@ -12,7 +12,7 @@ use stdext::function_name;
 pub async fn mk_lib_database_metadata_image_count(
     sqlx_pool: &sqlx::PgPool,
     class_id: i32,
-) -> Result<i32, sqlx::Error> {
+) -> Result<i64, sqlx::Error> {
     #[cfg(debug_assertions)]
     {
         mk_lib_logging::mk_logging_post_elk(
@@ -22,7 +22,7 @@ pub async fn mk_lib_database_metadata_image_count(
         .await
         .unwrap();
     }
-    let row: (i32,) = sqlx::query_as(
+    let row: (i64,) = sqlx::query_as(
         "select count(*) from mm_media \
         where mm_media_class_guid = $1",
     )
